Add AAIO Deployment file
Unused files are deleted for the time being but a service resource
will be needed in the future. Ingress is kept still since I tried
to reach the pod from the outside as well but might require some changes
too. Values.yaml is appended with the values from the default.yaml
so that helm lint wouldn't cry about it. Scorpio is now deployed in
the third order for easier deployment and isolation purposes

Signed-off-by: Meric Feyzullahoglu <meric.feyzullahoglu@gmail.com>